Start a new .Rmd file from a template.

Create a .Rmd file in a specified location with the basic layout provided, ready for the user to create a new question.

Usage

create_rmd(save_as, question_type, dynamic_type = "none", open = FALSE)

Arguments

save_as

Where to save the created file.

question_type

What kind of question being created. From:

"schoice"

"mchoice"

"string"

dynamic_type

How to make the question dynamic. Default is "none". From:

"none" - No code blocks are used for random question generation. Normal.

"variations" - Create a dataframe that contains full variations of a question.

"keywords" - Like "variations", but only a few key words in the question are changed between variations, as well as the answers.

"dynamic" - ("schoice" and "mchoice" only). Start with a tribble of potential answers and draw from it for each variation. Other answers from the tribble are used as incorrect choices.

"pooled" - ("schoice" and "mchoice" only). Single question where the answer choices are randomly selected.

open

Whether or not to open the file upon creation. Default: false.

Examples for Dynamic Types

"none" - Q: "Who developed the famous equation e = mc^2?", A: "Albert Einstein".

"variations" - Different questions referring to the properties of helium atoms that should still be kept in the same file. Q: "What is the atomic mass of helium?", A: "4". Q: "What state of matter is helium", A: "Gas".

"keywords" - Questions asking the atomic masses of different elements. Q: "What is the atomic mass of HELIUM?", A: "4". Q: "What is the atomic mass of OXYGEN?", A: "16".

"dynamic" - Questions asking which elements belong to different columns. Q: "Which element(s) can be found in column 18?", A: "He", "Ne", "Ar", "Kr", "Xe", "Rn", or "Og". Q: "Which element(s) can be found in column 16?", A: "O", "S", "Se", "Te", "Po", or "Lv".

"pooled" - Q: "Which element has the largest atomic radius?", A: "Francium" (the incorrect answers can be drawn from a pool containing all other elements).

string Questions

If the question_type is "string", the dynamic_type can only be "none", "keywords", or "variations".

Credits

Brighton Pauli, 2023